refactor(app): co-locate secondary panel stories with their components
The 8 story files lived in components/right-panel/ — residue of an
incomplete rename (4dab96328) — importing everything from
../secondary-panel/. Move them next to the components they cover and
rewrite the relative imports. All stories verified live (every import
resolves; no duplicates existed in secondary-panel/); Ladle's
src/**/*.stories.tsx glob picks them up unchanged.
